I added comments to shallow2d.c
I also renamed several of the variables in shallow2d.c. In particular,
I renamed fh, fuh, and fhv to FU_h, FU_hu, and FU_hv, respectively. This
better reflects the names in the physics on page 3 of the document. I
did the same thing for g.

Also, for each function, I added a short description of what the function
does as well as a short description of what each argument is/does.

I also added comments throughout the functions to give some insight into
what is going on.

Hopefully, with these comments, people will be able to read the code
and understand what's going on without having to spend an entire weekend
solving the jigsaw puzzle that was the original code.