I added some comments to central2d_correct_sd (I forgot to add them before, oops).
I also renamed a few variables (Ux -> U_x in central2d_correct... I
think this notation is closer to the standard notation for an x
partial derivative, which is what U_x holds).

I also added a few comments to code that calls complicated functions.
In particular, I added a comment to explain which argument a passed
valued corresponds to when it's not clear from the name of the variable
being passed... that description might make no sense. Look at line 553
in stepper.c... that's what I'm talking about.