I wrote most of the parallel code in stepper.
In particular, I wrote a new function to apply partitioned BC's.
To do this, I had to beef up the copy_subgrid function (accept a
stried for the source and dest).

I also added tons of openmp directives to the main while loop in
xrun. The code is not yet finished... I need to write code that writes
the boundary cells in each local U to global U.

After that, it's time to debug!